+++ src/VBox/VMM/PGMGst.h
@@ -297,7 +297,7 @@ static DECLCALLBACK(int) pgmR3Gst32BitWriteHandlerCR3(PVM pVM, RTGCPHYS GCPhys, 
          */
         const RTGCPTR   offPD = GCPhys & PAGE_OFFSET_MASK;
         const unsigned  iPD1  = offPD / sizeof(X86PDE);
-        const unsigned  iPD2  = (offPD + cbBuf - 1) / sizeof(X86PDE);
+        const unsigned  iPD2  = (unsigned)(offPD + cbBuf - 1) / sizeof(X86PDE);
         Assert(iPD1 - iPD2 <= 1);
         if (    (   pVM->pgm.s.pGuestPDR3->a[iPD1].n.u1Present
                  && pgmGetMapping(pVM, iPD1 << X86_PD_SHIFT) )
